scripts/requirements.txt
playwright
beautifulsoup4
python-dotenv
typer
tqdm
pandas
